requirements.txt
-e svn+http://code.djangoproject.com/svn/django/trunk#egg=Django-1.4a1-py2.7-dev
python-dateutil
